Factor out common code into toolcmd.py
