+++ tool_tests/rellic.py
@@ -3,15 +3,13 @@ import argparse
 import sys
 import shutil
 import logging
-import subprocess
-import signal
-from os import path, strerror
+from os import path
 import os
 from pathlib import Path
 from tqdm import tqdm
 from multiprocessing import Pool
 from functools import partial
-import re
+from toolcmd import ToolCmd
 
 log = logging.getLogger("rellic_test_suite")
 log.addHandler(logging.StreamHandler())
@@ -20,8 +18,6 @@ log.setLevel(logging.INFO)
 
 
 MYDIR = path.dirname(path.abspath(__file__))
-FILE_NAME_RE = re.compile("([^/\s]+\.[^/\s]+:\d+)")
-
 # given some input bitocode, run it through rellic and record outputs
 
 
@@ -54,25 +50,9 @@ FILE_NAME_RE = re.compile("([^/\s]+\.[^/\s]+:\d+)")
 #
 
 
-class RellicCmd:
-    def __init__(self, rellic, infile, outdir, source_base, index):
-        self.source_base = source_base
-        self.index = index
-        self.infile = infile
-        self.outdir = outdir
-        self.rellic = rellic
-        self.tmpout = None
-        self.cmd = self.make_rellic_cmd()
-        self.rc = None
-        self.out = None
-        self.err = None
+class RellicCmd(ToolCmd):
 
-    def set_output(self, rc, out, err):
-        self.rc = rc
-        self.out = out
-        self.err = err
-
-    def make_rellic_cmd(self):
+    def make_tool_cmd(self):
         f = self.infile.stem
         cfile = f"{self.index}-{f}.c"
         self.tmpout = self.outdir.joinpath("work").joinpath(cfile)
@@ -80,7 +60,7 @@ class RellicCmd:
         # rellic -logtostderr -input /input/dir/foo.bc -output /output/dir/work/foo.c
         log.debug(f"Setting tmpout to: {self.tmpout}")
         args = [
-            self.rellic,
+            self.tool,
             "--lower_switch",
             "--remove_phi_nodes",
             "-logtostderr",
@@ -90,40 +70,6 @@ class RellicCmd:
             str(self.tmpout),
         ]
         return args
-
-    def get_output_path(self):
-        rc_to_path = {
-            -131: "timeout",
-            -130: "oserror",
-            -129: "zero-sized-output",
-            -signal.SIGBUS: "sigbus",
-            -signal.SIGSEGV: "sigsegv",
-            -signal.SIGABRT: "sigabrt",
-            -signal.SIGILL: "sigill",
-            0: "success",
-        }
-
-        if self.rc != -signal.SIGABRT:
-            return rc_to_path.get(self.rc, "unknown")
-        else:
-            # First, check for a fatal error in the style of:
-            # F0415 05:22:54.866288 437680 IRToASTVisitor.cpp:123] Unknown LLVM Type
-            # Check only lines starting with 'F' since those are the fatal errors
-            for ln in self.err.splitlines():
-                if ln.startswith("F"):
-                    fname = FILE_NAME_RE.search(ln)
-                    if fname:
-                        return fname.group(0)
-
-            # Next, check for more generic filename matches in the whole message
-            # example:
-            # UNREACHABLE executed at /__w/cxx-common/cxx-common/vcpkg/buildtrees/llvm-11/src/org-11.0.0-8ebd641fb6.clean/llvm/lib/Support/APFloat.cpp:154!
-            fname = FILE_NAME_RE.search(self.err)
-            if fname:
-                return fname.group(0)
-            
-            # default to normal sigabrt handler
-            return rc_to_path.get(self.rc, "unknown")
 
     def save(self):
 
@@ -163,49 +109,6 @@ class RellicCmd:
             reprofile.write("#!/bin/sh\n")
             reprofile.write(" ".join(self.cmd))
             reprofile.write("\n")
-
-    def __del__(self):
-        if self.tmpout:
-            log.debug(f"Unlinking on delete {self.tmpout}")
-            try:
-                os.unlink(self.tmpout)
-            except FileNotFoundError as fnf:
-                log.debug(f"Tried to delete a file that doesn't exist: {self.tmpout}")
-
-    def run(self):
-
-        try:
-            log.debug(f"Running [{self.cmd}]")
-            rellic_run = subprocess.run(
-                self.cmd,
-                universal_newlines=True,
-                check=True,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                timeout=120, # two minutes should be more than enough
-            )
-        except OSError as oe:
-            log.debug("Rellic invocation hit OS error")
-            self.set_output(-130, "", oe.strerror)
-            return -130
-        except subprocess.CalledProcessError as cpe:
-            log.debug("Rellic invocation errored")
-            self.set_output(cpe.returncode, cpe.stdout, cpe.stderr)
-            return cpe.returncode
-        except subprocess.TimeoutExpired as tme:
-            log.debug("Rellic hit a timeout")
-            self.set_output(-131, tme.stdout, tme.stderr)
-            return -131
-
-        if 0 == os.path.getsize(self.tmpout):
-            self.set_output(
-                -129, rellic_run.stdout, rellic_run.stderr + "\n" + "Zero sized output"
-            )
-            return -129
-
-        # returncode should always be zero
-        self.set_output(rellic_run.returncode, rellic_run.stdout, rellic_run.stderr)
-        return rellic_run.returncode
 
 
 def run_rellic(rellic, output_dir, failonly, source_path, input_and_idx):
